pkg/aflow: move history formatting to llm_agent
History formatting and tag disarming are general agent utilities rather
than judge logic, and passing large payloads to fmt.Fprintf incurs
redundant allocations.
